requirement.txt
Flask==0.11.1
Flask-Compress==1.3.2
peewee==2.8.0
timeout-decorator==0.3.2
pyserial==3.1.1
RPi.GPIO==0.6.1
RPIO==0.10.0
mockito==0.7.0
